refactor: rename useAuth.js → AuthGate.js to preserve existing import paths
Avoids updating 38 consumer files — only main.jsx changes (imports
AuthGateProvider instead of AuthGate for the component).
