+++ app/ui/src/components/GovernancePage.jsx
@@ -1,5 +1,5 @@
 ﻿import { useState, useEffect, useCallback } from 'react';
-import { useAuth } from '../auth/useAuth';
+import { useAuth } from '../auth/AuthGate';
 import { formatDate } from '../utils/formatters';
 
 function formatNum(n) {
