+++ app/ui/src/components/SystemsPage.jsx
@@ -1,5 +1,5 @@
 ﻿import { useState, useEffect, useCallback } from 'react';
-import { useAuth } from '../auth/useAuth';
+import { useAuth } from '../auth/AuthGate';
 
 function formatRelativeTime(dateStr) {
   if (!dateStr) return 'Never';
