+++ app/ui/src/main.jsx
@@ -1,7 +1,7 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import './index.css'
-import AuthGate from './auth/AuthGate.jsx'
+import AuthGate from './auth/AuthGateProvider.jsx'
 import App from './App.jsx'
 
 createRoot(document.getElementById('root')).render(
